refactor: Table 타입 분리 및 Type네이밍 변경
