+++ src/page/LoginPage/components/LinkButton/LinkButton.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 
-export interface LinkButtonProps {
+interface LinkButtonProps {
   url: string;
   title: string;
 }
