+++ src/page/TableListPage/TableListPage.tsx
@@ -1,4 +1,5 @@
 import DefaultLayout from '../../layout/defaultLayout/DefaultLayout';
+import { DebateTable } from '../../type/type';
 import AddTable from './components/AddTable';
 import Table from './components/Table';
 
@@ -9,12 +10,6 @@ const data = [
   { name: '테이블 4', type: '의회식 토론', time: 30 },
 ];
 
-export interface TableProps {
-  name: string;
-  type: string;
-  time: number;
-}
-
 export default function TableListPage() {
   return (
     <DefaultLayout>
@@ -23,7 +18,7 @@ export default function TableListPage() {
       </DefaultLayout.Header>
       <div className="flex h-screen flex-col px-4 py-6">
         <main className="grid grid-cols-3 justify-items-center gap-6">
-          {data.map((table: TableProps, idx: number) => (
+          {data.map((table: DebateTable, idx: number) => (
             <Table
               key={idx}
               name={table.name}
